Move global hook tests into hooks test file
